+++ bob/build.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/benchkram/bob/pkg/env"
+	"github.com/benchkram/bob/pkg/envutil"
 	"github.com/benchkram/errz"
 
 	"github.com/benchkram/bob/bob/playbook"
@@ -28,7 +28,7 @@ func (b *B) Build(ctx context.Context, taskName string) (err error) {
 	b.PrintVersionCompatibility(ag)
 
 	for i, task := range ag.BTasks {
-		task.SetEnv(env.MergeEnv(task.Env(), b.env))
+		task.SetEnv(envutil.MergeEnv(task.Env(), b.env))
 		ag.BTasks[i] = task
 	}
